pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-unity-catalog-checkpoint"
version = "0.0.1"
description = "Unity Catalog-backed persistence for LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect>=17.0.10",
    "databricks-langchain>=0.9.0",
    "databricks-sdk>=0.73.0",
    "langchain>=1.0.3",
    "langgraph>=1.0.2",
    "langmem>=0.0.30",
    "loguru>=0.7.3",
    "mlflow==3.5.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.6.0",
    "twine>=5.0.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]  # Allow missing annotations and asserts in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --strict-markers --tb=short --color=yes"
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring actual Databricks connection",
    "slow: Tests that take a long time to run",
    "asyncio: Async tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
